V2.2 [1]app模块：    修改music列表与详情    增加category列表接口 [2]console模块：    修改music新增与更新接口，增加参数typeId    增加category的列表、详情、新增、修改、删除接口 [3]通过提升一层解决循环依赖    music新增与更新接口，增加校验typeId    优化分类删除逻辑 [4]list模糊查询升级    console使用联表查询    app使用子查询，利用${}替代子查询语句，抗高并发

+++ console/src/main/java/top/yuhanpeng/musiccard/console/controller/CategoryController.java
@@ -10,6 +10,7 @@ import top.yuhanpeng.musiccard.console.domain.CategoryListFeedVO;
 import top.yuhanpeng.musiccard.console.domain.CategoryListVO;
 import top.yuhanpeng.musiccard.module.entity.Category;
 import top.yuhanpeng.musiccard.module.service.CategoryService;
+import top.yuhanpeng.musiccard.module.service.MusicService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -27,6 +28,8 @@ import java.util.List;
 public class CategoryController {
     @Autowired
     private CategoryService categoryService;
+    @Autowired
+    private MusicService musicService;
 
     @RequestMapping("/category/list")
     public CategoryListFeedVO getCategoryList() {
@@ -99,12 +102,17 @@ public class CategoryController {
 
     @RequestMapping("/category/delete")
     public String categoryDelete(@RequestParam(value = "id") Long id) {
+        Long countMusicNum = musicService.getByTypeId(id);
+        if (countMusicNum > 0) {
+            log.error("删除失败，该分类下仍有音乐，无法删除");
+            return "删除失败，该分类下仍有音乐，无法删除";
+        }
         Integer affectedRows = 0;
         String res = "";
         try {
             affectedRows = categoryService.delete(id);
         } catch (Exception e) {
-            res = "无法找到id或者该分类下仍有音乐，无法删除";
+            res = "无法找到该id";
             log.error("cannot find the id", e);
         }
         if (affectedRows != 0) {
